--- libs/datalog/README.md
@@ -1,3 +1,0 @@
-# Datalog
-
-A tiny libraty to create CSV datalog files.
