--- libs/datalog/targetoverrides.ts
@@ -1,4 +1,0 @@
-// add target specific things here
-namespace datalog {
-    datalog.setStorage(new StorageDatalogStorage("datalog.csv"));
-}
