--- libs/datalog/test.ts
@@ -1,6 +1,0 @@
-let k = 0;
-forever(function () {
-    datalog.addRow()
-    datalog.addValue("x", k)
-    datalog.addValue("y", 1 / k)
-})
