+++ libs/datalogger/storagedatalog.ts
@@ -1,9 +1,9 @@
-namespace datalog {
+namespace datalogger {
     const FILENAME = "datalog.csv";
     /**
         * A storage for datalog data
     */
-    export class StorageDatalogStorage extends DatalogStorage {
+    export class FileStorage extends Storage {
         filename: string;
         constructor(filename: string) {
             super()
@@ -19,15 +19,15 @@ namespace datalog {
          * Appends the headers in datalog
          */
         appendHeaders(headers: string[]): void { 
-            const line = headers.join(datalog.SEPARATOR);
-            storage.appendLine(this.filename, `sep=${datalog.SEPARATOR}`);
+            const line = headers.join(datalogger.SEPARATOR);
+            storage.appendLine(this.filename, `sep=${datalogger.SEPARATOR}`);
             storage.appendLine(this.filename, line);
         }
         /**
          * Appends a row of data
          */
         appendRow(values: number[]): void { 
-            const line = values.join(datalog.SEPARATOR);
+            const line = values.join(datalogger.SEPARATOR);
             storage.appendLine(this.filename, line);
         }
         /**
