+++ libs/datalogger/targetoverrides.ts
@@ -1,0 +1,4 @@
+// add target specific things here
+namespace datalogger {
+    datalogger.setStorage(new FileStorage("log.csv"));
+}
